requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.10
python-dotenv==1.0.0
python-dateutil>=2.8.2